Pull block layer fix from Jan Kára.
Pull series.conf cleanup from Jiří Slabý.
Pull powerpc fix from Michal Suchánek.
Pull various fixes (git-fixes) from Oliver Neukum.

suse-commit: e26eecea396878e941455bae816bc2b3bbc14065

+++ drivers/net/wireless/mediatek/mt76/mt76x02.h
@@ -117,7 +117,7 @@ struct mt76x02_dev {
 
 	bool no_2ghz;
 
-	u8 coverage_class;
+	s16 coverage_class;
 	u8 slottime;
 
 	struct mt76x02_dfs_pattern_detector dfs_pd;
